refac(rfr): separate chunked reader and writer code
The code for chunked recordings was a bit of a mess. A lot of the code
was still sitting in `chunked/mod.rs`, This made it more complicated to
see what was visible and what depended on what.

This change moves the writer and reader specific code to their own
modules. The modules themselves are private and only those types which
need to be public are re-exported from the `chunked` module.

The writer code was moved to a new `chunked::write` module. This
includes `ChunkedWriter` and `ChunkBuffer` as well as their supporting
errors and notifiers.

The reader code was moved to a new `chunked::read` module. This includes
`Recording` and `Chunk` loader as well as thir supporting errors. The
method `Chunk::try_from_io` was moved into the `read` module as a free
function `read_chunk_from_io` as it's only used from the `ChunkLoader`.
